+++ src/textual/css/stylesheet.py
@@ -418,6 +418,7 @@ class Stylesheet:
         node: DOMNode,
         *,
         animate: bool = False,
+        cache: dict[tuple, RulesMap] | None = None,
     ) -> None:
         """Apply the stylesheet to a DOM node.
 
@@ -428,6 +429,7 @@ class Stylesheet:
                 classes modifying the same CSS property), then only the most specific
                 rule will be applied.
             animate: Animate changed rules.
+            cache: An optional cache when applying a group of nodes.
         """
         # Dictionary of rule attribute names e.g. "text_background" to list of tuples.
         # The tuples contain the rule specificity, and the value for that rule.
@@ -436,6 +438,23 @@ class Stylesheet:
         # same attribute, then we can choose the most specific rule and use that.
         rule_attributes: defaultdict[str, list[tuple[Specificity6, object]]]
         rule_attributes = defaultdict(list)
+
+        cache_key: tuple | None
+        if cache is not None:
+            cache_key = (
+                node._parent,
+                node._id,
+                node.classes,
+                node.pseudo_classes,
+                node._css_type_name,
+            )
+            cached_result: RulesMap | None = cache.get(cache_key)
+            if cached_result is not None:
+                self.replace_rules(node, cached_result, animate=animate)
+                self._process_component_classes(node)
+                return
+        else:
+            cache_key = None
 
         _check_rule = self._check_rule
         css_path_nodes = node.css_path_nodes
@@ -520,8 +539,18 @@ class Stylesheet:
                         rule_value = getattr(_DEFAULT_STYLES, initial_rule_name)
                     node_rules[initial_rule_name] = rule_value  # type: ignore[literal-required]
 
+            if cache is not None:
+                assert cache_key is not None
+                cache[cache_key] = node_rules
             self.replace_rules(node, node_rules, animate=animate)
+        self._process_component_classes(node)
 
+    def _process_component_classes(self, node: DOMNode) -> None:
+        """Process component classes for the given node.
+
+        Args:
+            node: A DOM Node.
+        """
         component_classes = node._get_component_classes()
         if component_classes:
             # Create virtual nodes that exist to extract styles
@@ -628,14 +657,15 @@ class Stylesheet:
             nodes: Nodes to update.
             animate: Enable CSS animation.
         """
+        cache: dict[tuple, RulesMap] = {}
         apply = self.apply
 
         for node in nodes:
-            apply(node, animate=animate)
+            apply(node, animate=animate, cache=cache)
             if isinstance(node, Widget) and node.is_scrollable:
                 if node.show_vertical_scrollbar:
-                    apply(node.vertical_scrollbar)
+                    apply(node.vertical_scrollbar, cache=cache)
                 if node.show_horizontal_scrollbar:
-                    apply(node.horizontal_scrollbar)
+                    apply(node.horizontal_scrollbar, cache=cache)
                 if node.show_horizontal_scrollbar and node.show_vertical_scrollbar:
-                    apply(node.scrollbar_corner)
+                    apply(node.scrollbar_corner, cache=cache)